+++ app/Api/Uri.php
@@ -183,6 +183,20 @@ class Uri
     }
 
     #[ArrayShape(['uri' => "string", 'name' => "string"])]
+    public static function resource(string $resource_type_id, string $resource_id, array $parameters = []): array
+    {
+        $uri = '/' . self::VERSION . '/resource-types/' . $resource_type_id . '/resources/' . $resource_id;
+        if (count($parameters) > 0) {
+            $uri .= '?' . http_build_query($parameters);
+        }
+
+        return [
+            'uri' => $uri,
+            'name' => 'Resource'
+        ];
+    }
+
+    #[ArrayShape(['uri' => "string", 'name' => "string"])]
     public static function resources(string $resource_type_id, array $parameters = []): array
     {
         $uri = '/' . self::VERSION . '/resource-types/' . $resource_type_id . '/resources';