+++ app/Http/Controllers/Share.php
@@ -6,7 +6,10 @@ namespace App\Http\Controllers;
 use App\Actions\Game\Log;
 use App\Api\Service;
 use App\Models\ShareToken;
+use App\Notifications\ApiError;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Config;
+use Illuminate\Support\Facades\Notification;
 use JetBrains\PhpStorm\ArrayShape;
 
 /**
@@ -180,7 +183,13 @@ class Share extends Controller
         );
 
         if ($log_action_result !== 201) {
-            // @todo - Log an error
+            $config = Config::get('app.config');
+
+            Notification::route('mail', $config['error_email'])
+                ->notify(new ApiError(
+                    'Unable to log the score for the upper section',
+                    $log_action->getMessage()
+                ));
         }
 
         return $this->score(
@@ -236,7 +245,13 @@ class Share extends Controller
         );
 
         if ($log_action_result !== 201) {
-            // @todo - Log an error
+            $config = Config::get('app.config');
+
+            Notification::route('mail', $config['error_email'])
+                ->notify(new ApiError(
+                    'Unable to log the score for the lower section',
+                    $log_action->getMessage()
+                ));
         }
 
         return $this->score(