+++ app/Models/PartialRegistration.php
@@ -1,0 +1,20 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Models;
+
+use Illuminate\Database\Eloquent\Model;
+
+/**
+ * @author Dean Blackborough <dean@g3d-development.com>
+ * @copyright Dean Blackborough (Costs to Expect) 2018-2022
+ * https://github.com/costs-to-expect/yahtzee/blob/main/LICENSE
+ *
+ * @property string $token
+ * @property string $email
+ */
+class PartialRegistration extends Model
+{
+    protected $table = 'partial_registration';
+}