+++ database/migrations/2022_08_25_140046_create_partial_registration_table.php
@@ -1,0 +1,26 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+return new class extends Migration
+{
+    public function up()
+    {
+        Schema::create('partial_registration', function (Blueprint $table) {
+
+            $table->charset = 'utf8mb4';
+            $table->collation = 'utf8mb4_unicode_ci';
+
+            $table->string('token')->primary();
+            $table->string('email');
+            $table->timestamps();
+        });
+    }
+
+    public function down()
+    {
+        // No down
+    }
+};