+++ resources/views/components/offcanvas.blade.php
@@ -1,8 +1,16 @@
 <nav class="navbar navbar-dark bg-dark" aria-label="Offcanvas navbar">
     <div class="container-fluid">
-        <a class="navbar-brand" href="/">
+        @auth
+        <a class="navbar-brand" href="{{ route('home') }}">
             Yahtzee by <img src="{{ asset('images/logo.png') }}" width="30" height="30" class="d-inline-block align-middle" alt=""><span class="d-none">C</span>osts to Expect.com
         </a>
+        @endauth
+
+        @guest
+            <a class="navbar-brand" href="/">
+                Yahtzee by <img src="{{ asset('images/logo.png') }}" width="30" height="30" class="d-inline-block align-middle" alt=""><span class="d-none">C</span>osts to Expect.com
+            </a>
+        @endguest
         <button class="navbar-toggler" type="button" data-bs-toggle="offcanvas" data-bs-target="#offcanvasNavbarDark" aria-controls="offcanvasNavbarDark">
             <span class="navbar-toggler-icon"></span>
         </button>
@@ -30,6 +38,9 @@
                     </li>
                     <li class="nav-item">
                         <a class="nav-link" href="https://api.costs-to-expect.com">The API</a>
+                    </li>
+                    <li class="nav-item">
+                        <a class="nav-link @if($active === 'account') active @endif" @if($active === 'account') aria-current="page" @endif href="{{ route('account') }}">Account</a>
                     </li>
                     <li class="nav-item">
                         <a class="nav-link" href="{{ route('sign-out') }}">Sign-out</a>